This is workikng very well, thanks! I have one more request. The current file naming logic is this:
If a filename contains "EMR" or "Service" (case-insensitive), that file's name will be used as the base for the output Excel file.
I'll remove "Audit" (case-insensitive) from this base name if it exists.
Then, I'll remove the original extension (.txt or .pipe) and append .xlsx.
If multiple files match (e.g., one "EMR" and one "Service" file), the first one found in the list will dictate the name.
If no file contains "EMR" or "Service", the output file will be named "Combined_Output.xlsx" as it is currently.

I'd like to change it to this:

If a filename contains "EMR" or "Service" (case-insensitive), that file's name will be used as the base for the output Excel file.
I'll remove "Audit" (case-insensitive) from this base name if it exists.
Then, I'll remove the original extension (.txt or .pipe) and append .xlsx.
If multiple files match (e.g., one "EMR" and one "Service" file), the first one found in the list will dictate the name.
If no file contains "EMR" or "Service", the output file will be named after the first pipe delimited file ingested.